+++ src/ic/url.mli
@@ -4,10 +4,12 @@ type parsed =
   | Ic of string
   | IcAlias of string
   | FileValue of string
+  | IdlFile of string
   | Prim
 
 val string_of_parsed : parsed -> string
 val parse : string -> (parsed, string) result
+val idl_extension : string
 val idl_basename_of_blob : string -> string
 val idl_basename_of_envvar : string -> string