+++ src/idllib/escape.mli
@@ -1,5 +1,6 @@
 (* This (or a type like this) could move into the IDL AST *)
 type label = Nat of Lib.Uint32.t | Id of string
+val pascal_case : string -> string
 val escape : string -> string
 val escape_num : Lib.Uint32.t -> string
 val escape_method : Source.region -> string -> string